+++ core/src/main/java/org/hisrc/zugradarscraper/model/Point.java
@@ -1,16 +1,15 @@
 package org.hisrc.zugradarscraper.model;
 
-import java.math.BigDecimal;
 
 import org.apache.commons.lang3.Validate;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class Point extends Geometry<BigDecimal[]> {
+public class Point extends Geometry<double[]> {
 
 	@JsonCreator
-	public Point(@JsonProperty("coordinates") BigDecimal[] coordinates) {
+	public Point(@JsonProperty("coordinates") double[] coordinates) {
 		super(coordinates);
 		Validate.isTrue(coordinates.length >= 2);
 	}
@@ -20,6 +19,6 @@ public class Point extends Geometry<BigDecimal[]> {
 //		Validate.isTrue(points.size() > 1);
 //		double lon = points.stream().map(Point::getCoordinates).collect(Collectors.averagingDouble(c -> c[0]));
 //		double lat = points.stream().map(Point::getCoordinates).collect(Collectors.averagingDouble(c -> c[1]));
-//		return new Point(new BigDecimal[] { lon, lat });
+//		return new Point(new double[] { lon, lat });
 //	}
 }
