+++ core/src/main/java/org/hisrc/zugradarscraper/model/TrainRouteMultiLineStringFeatureCollection.java
@@ -1,10 +1,9 @@
 package org.hisrc.zugradarscraper.model;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public class TrainRouteMultiLineStringFeatureCollection extends
-		FeatureCollection<TrainRouteMultiLineStringFeature, MultiLineString, BigDecimal[][][], TrainRouteMultiLineStringFeature.Properties> {
+		FeatureCollection<TrainRouteMultiLineStringFeature, MultiLineString, double[][][], TrainRouteMultiLineStringFeature.Properties> {
 
 	public TrainRouteMultiLineStringFeatureCollection(List<TrainRouteMultiLineStringFeature> features) {
 		super(features);
